VIX-1046 - Mark Collections - Disable the delete button when no Mark Collection is selected in the docker. Remove un need code due to this change.

+++ Modules/Editor/TimedSequenceEditor/Form_Marks.cs
@@ -2,8 +2,6 @@
 using System.Linq;
 using System.Windows.Forms;
 using System.Drawing;
-using System.Collections.Generic;
-using System.Windows.Forms.VisualStyles;
 using Common.Controls;
 using Common.Controls.Scaling;
 using Common.Controls.Theme;
@@ -36,6 +34,7 @@ namespace VixenModules.Editor.TimedSequenceEditor
 			ThemeUpdateControls.UpdateControls(this);
 			contextMenuStrip1.Renderer = new ThemeToolStripRenderer();
 			listViewMarkCollections.BackColor = ThemeColorTable.BackgroundColor; //Over-rides the default Listview background
+			toolStripButtonDeleteMarkCollection.Enabled = false;
 		}
 
 		private void Form_Marks_Load(object sender, EventArgs e)
@@ -180,7 +179,7 @@ namespace VixenModules.Editor.TimedSequenceEditor
 
 		private void listViewMarkCollections_PreviewKeyDown(object sender, PreviewKeyDownEventArgs e)
 		{
-			if (e.KeyCode == Keys.Delete)
+			if (e.KeyCode == Keys.Delete && listViewMarkCollections.SelectedItems.Count > 0)
 			{
 				DeleteSelectedMarkCollections();
 			}
@@ -188,29 +187,19 @@ namespace VixenModules.Editor.TimedSequenceEditor
 
 		private void DeleteSelectedMarkCollections()
 		{
-			if (listViewMarkCollections.SelectedItems.Count > 0)
+			//messageBox Arguments are (Text, Title, No Button Visible, Cancel Button Visible)
+			MessageBoxForm.msgIcon = SystemIcons.Question; //adds a system icon to the message form.
+			var messageBox = new MessageBoxForm("Are you sure you want to delete the selected Marks in the Collection?", "Delete Mark Collection", true, false);
+			messageBox.ShowDialog();
+			if (messageBox.DialogResult == DialogResult.OK)
 			{
-				//messageBox Arguments are (Text, Title, No Button Visible, Cancel Button Visible)
-				MessageBoxForm.msgIcon = SystemIcons.Question; //adds a system icon to the message form.
-				var messageBox = new MessageBoxForm("Are you sure you want to delete the selected Marks in the Collection?", "Delete Mark Collection", true, false);
-				messageBox.ShowDialog();
-				if (messageBox.DialogResult == DialogResult.OK)
+				foreach (ListViewItem item in listViewMarkCollections.SelectedItems)
 				{
-					foreach (ListViewItem item in listViewMarkCollections.SelectedItems)
-					{
-						listViewMarkCollections.SelectedItems[0].Remove();
-						MarkCollection mc = item.Tag as MarkCollection;
-						Sequence.MarkCollections.Remove(mc);
-					}
-					OnChangedMarkCollection(new MarkCollectionArgs(null));
+					listViewMarkCollections.SelectedItems[0].Remove();
+					MarkCollection mc = item.Tag as MarkCollection;
+					Sequence.MarkCollections.Remove(mc);
 				}
-			}
-			else
-			{
-				//messageBox Arguments are (Text, Title, No Button Visible, Cancel Button Visible)
-				MessageBoxForm.msgIcon = SystemIcons.Error; //adds a system icon to the message form.
-				var messageBox = new MessageBoxForm("Please select a Mark Collection to delete and press the delete button again.", "Delete Mark Collection", false, false);
-				messageBox.ShowDialog();
+				OnChangedMarkCollection(new MarkCollectionArgs(null));
 			}
 		}
 
@@ -280,6 +269,11 @@ namespace VixenModules.Editor.TimedSequenceEditor
 				}
 				OnMarkCollectionChecked(new MarkCollectionArgs(mc));
 			}
+		}
+
+		private void listViewMarkCollections_SelectedIndexChanged(object sender, EventArgs e)
+		{
+			toolStripButtonDeleteMarkCollection.Enabled = listViewMarkCollections.SelectedItems.Count > 0;
 		}
 
 	}
